update configuration strategy
* a configuration file can be passed as first argument, resolved in the working directory
* a default configuration is loaded if no one is passed
* an example configuration file is available to showcase everything that can be done
* a bin 'access-watch' is create and configured in the package.json
