+++ bin/access-watch
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../index')
