+++ config/example.js
@@ -1,0 +1,164 @@
+const pipeline = require('../lib/pipeline')
+
+/* Input configuration
+====================== */
+
+const input = require('../input')
+const format = require('../format')
+
+/* Syslog inputs
+---------------- */
+
+/* Syslog input in Nginx 'combined' format */
+
+// const syslogNginxCombinedInput = input.syslog.create({
+//   name: 'Syslog (nginx combined format)',
+//   port: 1514,
+//   parse: format.nginx.parser({format: format.nginx.formats.combined})
+// })
+
+// pipeline.registerInput(syslogNginxCombinedInput)
+
+/* Syslog input in Nginx 'access_watch' format */
+
+// const syslogNginxAccessWatchInput = input.syslog.create({
+//   name: 'Syslog (nginx access_watch format)',
+//   port: 1515,
+//   parse: format.nginx.parser({format: format.nginx.formats.accessWatch})
+// })
+
+// pipeline.registerInput(syslogNginxAccessWatchInput)
+
+const syslogInput = input.syslog.create({
+  name: 'Syslog (JSON standard format)',
+  port: 1516,
+  parse: format.json.parser()
+})
+
+pipeline.registerInput(syslogInput)
+
+/* HTTP inputs
+-------------- */
+
+const httpInput = input.http.create({
+  name: 'HTTP server (JSON standard format)',
+  path: '/input/log',
+  parse: format.json.parser()
+})
+
+pipeline.registerInput(httpInput)
+
+/* WebSocket inputs
+------------------- */
+
+/* WebSocket server input in standard JSON format (listening for logs) */
+
+// const webSocketServerInput = input.websocket.create({
+//   name: 'WebSocket server (JSON standard format)',
+//   type: 'server',
+//   path: '/input/log',
+//   parse: format.json.parser()
+// })
+
+// pipeline.registerInput(webSocketServerInput)
+
+/* WebSocket client input in standard JSON format (subscribing to logs) */
+
+// const websocketClientInput = input.websocket.create({
+//  address: 'ws://HOST:PORT',
+//  parse: format.json.parser()
+// })
+
+// pipeline.registerInput(websocketClientInput)
+
+/* File inputs
+-------------- */
+
+/* File inputs accepting Nginx format */
+
+// const fileInputNginxCombined = input.file.create({
+//   path: '/var/log/nginx/access.log',
+//   parse: format.nginx.parser({
+//     format: format.nginx.formats.combined
+//   })
+// })
+
+// pipeline.registerInput(fileInputNginxCombined)
+
+// const fileInputNginxAccessWatch = input.file.create({
+//   path: '/var/log/nginx/blogmarks_access_watch.log',
+//   parse: format.nginx.parser({
+//     format: format.nginx.formats.accessWatch
+//   })
+// })
+
+// pipeline.registerInput(fileInputNginxAccessWatch)
+
+/* Elasticsearch inputs
+----------------------- */
+
+/* Elasticsearch input (polling) accepting a commonly used Logstash format */
+
+// const elasticsearchInput = input.elasticsearch.create({
+//   config: {
+//     host: '__HOST__:__PORT__'
+//   },
+//   query: {
+//     index: '__INDEX__',
+//     type: '__TYPE__'
+//   },
+//   parse: format.logstash.formats['HTTPD_COMBINEDLOG']
+// })
+
+// pipeline.registerInput(elasticsearchInput)
+
+/*
+Pipeline configuration
+======================
+*/
+
+const proxy = require('../plugins/proxy')
+const hub = require('../plugins/hub')
+const { selectKeys } = require('../lib/util')
+
+let stream = pipeline
+
+  /* Filter requests */
+
+  // .filter(log => log.getIn(['request', 'host']) === 'example.com')
+
+  /* Detect the public IP address if it's behind a proxy */
+
+  .map(log => log.set('address',
+                        proxy.detectAddress(log.getIn(['request', 'address']),
+                                            log.getIn(['request', 'headers']))))
+
+  /* Augment with data from the Access Watch Hub */
+
+  .map(log => {
+    return hub.fetchIdentity(Map({
+      address: log.getIn(['address', 'value']),
+      headers: log.getIn(['request', 'headers']),
+      captured_headers: log.getIn(['request', 'captured_headers'])
+    })).then(identity => {
+      if (identity) {
+        // Add identity properties
+        log = log.set('identity', selectKeys(identity, ['id', 'type', 'label']))
+        // Add identity objects
+        ;['address', 'user_agent', 'robot', 'reputation'].forEach(key => {
+          if (identity.has(key)) {
+            log = log.set(key, identity.get(key))
+          }
+        })
+      }
+      return log
+    })
+  })
+
+  /* Output to the console as JS object */
+
+  // .map(log => console.log(log.toJS()))
+
+module.exports = {
+  stream
+}
