+++ config/index.js
@@ -1,3 +1,13 @@
-require('./input')
+const path = require('path')
 
-require('./pipeline')
+let config
+
+if (process.argv[2]) {
+  config = path.resolve(process.cwd(), process.argv[2])
+} else {
+  config = path.resolve(__dirname, './default')
+}
+
+require(config)
+
+require('../dashboard')
